refactor: CLIN-716 change occurence for occurrence
