+++ src/views/screens/variant/OccurrenceDrawer/HcDescription.tsx
