+++ src/views/screens/variant/OccurrenceDrawer/IGVModal/index.module.scss
