+++ src/views/screens/variant/OccurrenceDrawer/index.module.scss
